renaming Code2 to EditorCodeOverlay
