--- components/Code2/index.js
@@ -1,2 +1,0 @@
-import Code2 from "./Code2.tsx";
-export default Code2;
