+++ components/EditorCodeOverlay/index.js
@@ -1,0 +1,2 @@
+import EditorCodeOverlay from "./EditorCodeOverlay.tsx";
+export default EditorCodeOverlay;
